requirements.txt
requests>=2.31.0
pathlib2>=2.3.7; python_version < '3.4'
python-dotenv>=1.0.0
resend>=0.6.0
packaging>=21.0
